chore: rename tests folder to e2eTests
