+++ tests/testsCore/src/androidTest/kotlin/com/wire/android/tests/core/e2eTests/ApplockTest.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
-package com.wire.android.tests.core.tests
+package com.wire.android.tests.core.e2eTests
 
 import android.content.Context
 import androidx.test.ext.junit.runners.AndroidJUnit4
